pyproject.toml
[project]
name = "databricks-gdpr-compliance"
version = "1.0.0"
description = "GDPR-compliant data governance framework for Databricks"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "databricks-sdk>=0.20.0",
    "databricks-sql-connector>=3.0.0",
    "pyspark>=3.5.0",
    "cryptography>=42.0.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "faker>=20.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

